Refactor CTC variables, mostly moving variables into their own folders
