+++ openfisca_us/variables/gov/irs/credits/ctc/ctc_reduction.py
@@ -1,8 +1,8 @@
 from openfisca_us.model_api import *
-from openfisca_us.variables.gov.irs.credits.ctc.maximum.individual.child import (
+from openfisca_us.variables.gov.irs.credits.ctc.maximum.individual.ctc_child_individual_maximum import (
     ctc_child_individual_maximum,
 )
-from openfisca_us.variables.gov.irs.credits.ctc.maximum.individual.adult import (
+from openfisca_us.variables.gov.irs.credits.ctc.maximum.individual.ctc_adult_individual_maximum import (
     ctc_adult_individual_maximum,
 )
 from openfisca_us.parameters import default_parameters
