+++ openfisca_us/variables/gov/irs/credits/ctc/maximum/ctc_maximum.py
@@ -9,5 +9,4 @@ class ctc_maximum(Variable):
     documentation = "Maximum value of the Child Tax Credit, before phaseout."
     definition_period = YEAR
 
-    def formula(tax_unit, period, parameters):
-        return aggr(tax_unit, period, ["ctc_individual_maximum"])
+    formula = sum_of_variables(["ctc_individual_maximum"])
