+++ openfisca_us/variables/gov/irs/credits/ctc/maximum/individual/ctc_child_individual_maximum.py
@@ -16,14 +16,12 @@ class ctc_child_individual_maximum(Variable):
 
     def formula(person, period, parameters):
         ctc = parameters(period).irs.credits.ctc
-        age = person("age", period)
-        is_child = age <= ctc.child.max_age
-        return is_child * ctc.child.amount
+        return person("is_ctc_qualifying_child", period) * ctc.child.amount
 
     def formula_2021(person, period, parameters):
         ctc = parameters(period).irs.credits.ctc
         age = person("age", period)
-        is_child = age <= ctc.child.max_age
+        is_child = person("is_ctc_qualifying_child", period)
         is_young_child = age <= ctc.child.young.max_age
         return (
             is_young_child * ctc.child.young.increase
